requirements.txt
notify-py
openai-whisper
sounddevice
wavio
ollama